+++ wire-ios/Wire-iOS/Sources/Developer/DeveloperTools/E2eIdentity/DeveloperDeviceDetailsSettingsSelectionViewModel.swift
@@ -17,6 +17,7 @@
 //
 
 import Foundation
+import WireDataModel
 
 // TODO: Remove this
 
@@ -99,6 +100,35 @@ class DeveloperDeviceDetailsSettingsSelectionViewModel: ObservableObject {
         case let .itemTapped(item):
             selectedItemID = item.id
             Self.selectedE2eIdentiyStatus = item.value
+        }
+    }
+
+    static func e2eIdentityProvider() -> E2eIdentityProviding {
+       if DeveloperDeviceDetailsSettingsSelectionViewModel.isE2eIdentityViewEnabled {
+           let status = E2EIdentityCertificateStatus.status(
+               for: DeveloperDeviceDetailsSettingsSelectionViewModel.selectedE2eIdentiyStatus ?? ""
+           )
+           switch status {
+           case .notActivated:
+               return MockNotActivatedE2eIdentityProvider()
+           case .revoked:
+               return MockRevokedE2eIdentityProvider()
+           case .expired:
+               return MockExpiredE2eIdentityProvider()
+           case .valid:
+               return MockValidE2eIdentityProvider()
+           case .none:
+               return E2eIdentityProvider()
+           }
+       }
+       return E2eIdentityProvider()
+   }
+
+    static func mlsProvider() -> MLSProviding {
+        if DeveloperFlag.enableMLSSupport.isOn {
+            MockMLSProvider(isMLSEnbaled: DeveloperFlag.enableMLSSupport.isOn)
+        } else {
+            MLSProvider()
         }
     }
 }
